+++ src/Dosepy/old/gui_components/Bloque_gamma.py
@@ -14,12 +14,12 @@ from PyQt6.QtCore import Qt
 from PyQt6.QtGui import QIcon
 from matplotlib.figure import Figure
 import numpy as np
-from Dosepy.GUILayouts.Bloque_Imagenes import Qt_Figure_Imagen
+from .Bloque_Imagenes import Qt_Figure_Imagen
 
 #from GUILayouts.Bloque_Imagenes import Qt_Figure_Imagen
 from importlib import resources
 
-import Dosepy.dose as dp
+import Dosepy.old.dose as dp
 
 
 class Bloque_gamma(QWidget):
