+++ src/Dosepy/old/gui_components/licencia_window.py
@@ -15,9 +15,9 @@
 import sys
 from importlib import resources
 
-from PySide6.QtWidgets import QWidget, QLabel, QVBoxLayout, QHBoxLayout, QApplication, QPlainTextEdit
-from PySide6.QtGui import QPixmap
-from PySide6.QtCore import Qt
+from PyQt6.QtWidgets import QWidget, QLabel, QVBoxLayout, QHBoxLayout, QApplication, QPlainTextEdit
+from PyQt6.QtGui import QPixmap
+from PyQt6.QtCore import Qt
 
 
 class Licencia_Window(QWidget):
